Move solana proto definitions and conversion methods to new crate
