+++ ledger/Cargo.toml
@@ -44,6 +44,7 @@ solana-runtime = { path = "../runtime", version = "1.5.0" }
 solana-sdk = { path = "../sdk", version = "1.5.0" }
 solana-stake-program = { path = "../programs/stake", version = "1.5.0" }
 solana-storage-bigtable = { path = "../storage-bigtable", version = "1.5.0" }
+solana-storage-proto = { path = "../storage-proto", version = "1.5.0" }
 solana-vote-program = { path = "../programs/vote", version = "1.5.0" }
 tempfile = "3.1.0"
 thiserror = "1.0"
