+++ storage-bigtable/src/lib.rs
@@ -2,15 +2,15 @@ use log::*;
 use serde::{Deserialize, Serialize};
 use solana_sdk::{
     clock::{Slot, UnixTimestamp},
-    deserialize_utils::default_on_eof,
     pubkey::Pubkey,
     signature::Signature,
     sysvar::is_sysvar_id,
     transaction::{Transaction, TransactionError},
 };
+use solana_storage_proto::convert::generated;
 use solana_transaction_status::{
     ConfirmedBlock, ConfirmedTransaction, ConfirmedTransactionStatusWithSignature, Reward,
-    RewardType, TransactionStatus, TransactionStatusMeta, TransactionWithStatusMeta,
+    TransactionStatus, TransactionStatusMeta, TransactionWithStatusMeta,
 };
 use std::{collections::HashMap, convert::TryInto};
 use thiserror::Error;
@@ -21,10 +21,7 @@ extern crate serde_derive;
 mod access_token;
 mod bigtable;
 mod compression;
-pub mod convert;
 mod root_ca_certificate;
-
-use convert::generated;
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -234,53 +231,6 @@ impl From<Reward> for StoredConfirmedBlockReward {
             pubkey, lamports, ..
         } = value;
         Self { pubkey, lamports }
-    }
-}
-
-pub type StoredExtendedRewards = Vec<StoredExtendedReward>;
-
-#[derive(Serialize, Deserialize)]
-pub struct StoredExtendedReward {
-    pubkey: String,
-    lamports: i64,
-    #[serde(deserialize_with = "default_on_eof")]
-    post_balance: u64,
-    #[serde(deserialize_with = "default_on_eof")]
-    reward_type: Option<RewardType>,
-}
-
-impl From<StoredExtendedReward> for Reward {
-    fn from(value: StoredExtendedReward) -> Self {
-        let StoredExtendedReward {
-            pubkey,
-            lamports,
-            post_balance,
-            reward_type,
-        } = value;
-        Self {
-            pubkey,
-            lamports,
-            post_balance,
-            reward_type,
-        }
-    }
-}
-
-impl From<Reward> for StoredExtendedReward {
-    fn from(value: Reward) -> Self {
-        let Reward {
-            pubkey,
-            lamports,
-            post_balance,
-            reward_type,
-            ..
-        } = value;
-        Self {
-            pubkey,
-            lamports,
-            post_balance,
-            reward_type,
-        }
     }
 }
