+++ storage-proto/build-proto/.gitignore
@@ -1,0 +1,1 @@
+target/
