+++ storage-proto/build-proto/src/main.rs
@@ -2,9 +2,9 @@ fn main() -> Result<(), std::io::Error> {
     let manifest_dir = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"));
 
     let out_dir = manifest_dir.join("../proto");
-    let googleapis = manifest_dir.join("googleapis");
+    let proto_files = manifest_dir.join("../src");
 
-    println!("Google API directory: {}", googleapis.display());
+    println!("Protobuf directory: {}", proto_files.display());
     println!("output directory: {}", out_dir.display());
 
     tonic_build::configure()
@@ -12,8 +12,5 @@ fn main() -> Result<(), std::io::Error> {
         .build_server(false)
         .format(true)
         .out_dir(&out_dir)
-        .compile(
-            &[googleapis.join("google/bigtable/v2/bigtable.proto")],
-            &[googleapis],
-        )
+        .compile(&[proto_files.join("confirmed_block.proto")], &[proto_files])
 }
