+++ storage-proto/src/confirmed_block.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package solana.bigtable.ConfirmedBlock;
+package solana.storage.ConfirmedBlock;
 
 message ConfirmedBlock {
     string previous_blockhash = 1;
